requirements.txt
streamlit
uvicorn
fastapi
python-dotenv
langchain_groq
langgraph
ipython
langchain_community
feedparser
sentence_transformers